autogen_lab/projects/meal_prep_rag_pipeline/requirements.txt
pandas-to-pydantic